cansei de escrever aqui

+++ sys/api/extra/patrocinadores/remove.php
@@ -31,6 +31,10 @@ $img = decrypt($img);
 
 $caminho = "../../../../imagens/patrocinadores";
 
-unlink("$caminho/$img");
-mysqli_query($__CONEXAO__, "delete from patrocinadores where id='$id'");
-endCode("Imagem excluida", true);
+if(unlink("$caminho/$img")){
+    mysqli_query($__CONEXAO__, "delete from patrocinadores where id='$id'");
+    endCode("Imagem excluida", true);
+} else {
+    endCode("Erro ao excluir imagem / caminho/$img", false);
+
+};
